refactor: organizes SkeletonTable exports in their respective folder
