+++ src/components/ui/skeleton-table/index.ts
@@ -1,0 +1,2 @@
+export { default as SkeletonTable } from './SkeletonTable.vue';
+export type { SkeletonTableProps } from './types';
