+++ src/components/ui/skeleton-table/types.ts
@@ -1,0 +1,11 @@
+export interface SkeletonTableProps {
+  cols?: string | number;
+  hideHeader?: boolean;
+  lines?: string | number;
+  noBorder?: boolean;
+  padding?: string;
+  rows?: string | number;
+  width?: string;
+  withAction?: boolean | string | number;
+  withAvatar?: boolean;
+}
